pyproject.toml
[project]
name = "clay-webhook-os"
version = "1.0.0"
description = "Clay Webhook OS — AI-powered webhook server for Clay HTTP Actions"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "pyyaml>=6.0.2",
    "httpx>=0.28.0",
    "parallel-web>=0.4.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.28.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
